+++ src/camera.rs
@@ -6,7 +6,7 @@ use crate::shaders::{
     pixelate::{PixelatePlugin, PixelateSettings},
 };
 
-use crate::{ship::Ship, state::AppState};
+use crate::{ship::Ship, state::GameState};
 
 pub struct CameraPlugin;
 impl Plugin for CameraPlugin {
@@ -14,7 +14,7 @@ impl Plugin for CameraPlugin {
         // app.add_plugins(PixelatePlugin);
         // app.add_plugins(ChromaticAberrationPlugin);
         app.add_systems(Startup, setup);
-        app.add_systems(Update, follow_player.run_if(in_state(AppState::Active)));
+        app.add_systems(Update, follow_player.run_if(in_state(GameState::Active)));
     }
 }