+++ src/hud/mod.rs
@@ -1,7 +1,7 @@
 use bevy::prelude::*;
 use bevy_rapier2d::prelude::Velocity;
 
-use crate::{assets::UiAssets, ship::Ship, state::AppState};
+use crate::{assets::UiAssets, ship::Ship, state::GameState};
 
 pub mod indicator;
 
@@ -15,8 +15,8 @@ pub struct HudPlugin;
 impl Plugin for HudPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugins(IndicatorPlugin);
-        app.add_systems(OnEnter(AppState::GameCreate), setup);
-        app.add_systems(Update, hud_speedometer.run_if(in_state(AppState::Active)));
+        app.add_systems(OnEnter(GameState::GameCreate), setup);
+        app.add_systems(Update, hud_speedometer.run_if(in_state(GameState::Active)));
     }
 }