+++ src/main.rs
@@ -1,4 +1,5 @@
-use bevy::{audio::PlaybackMode, prelude::*};
+use bevy::prelude::*;
+use bevy_asset_loader::prelude::*;
 use bevy_inspector_egui::quick::WorldInspectorPlugin;
 use bevy_rapier2d::prelude::*;
 use leafwing_input_manager::prelude::*;
@@ -21,7 +22,7 @@ mod star;
 mod state;
 
 use crate::{
-    assets::{AssetsPlugin, AudioAssets},
+    assets::{AudioAssets, SpriteAssets, UiAssets},
     background::BackgroundPlugin,
     camera::CameraPlugin,
     credits::CreditsPlugin,
@@ -32,13 +33,13 @@ use crate::{
     pause::PausePlugin,
     planetary_system::PlanetarySystemPlugin,
     ship::ShipPlugin,
-    state::{AppState, ForState, StatePlugin},
+    state::{GameState, StatePlugin},
 };
 
 fn main() {
     let mut app = App::new();
 
-    app.add_state::<AppState>();
+    app.add_state::<GameState>();
 
     app.add_plugins((
         DefaultPlugins
@@ -51,7 +52,6 @@ fn main() {
             })
             .set(ImagePlugin::default_nearest()),
         GameTimePlugin,
-        AssetsPlugin,
         RapierPhysicsPlugin::<NoUserData>::pixels_per_meter(1.0),
         InputManagerPlugin::<MenuAction>::default(),
         BackgroundPlugin,
@@ -72,35 +72,27 @@ fn main() {
         WorldInspectorPlugin::new(),
     ));
 
-    // app.insert_resource(ClearColor(Color::rgb(0., 0., 0.)));
+    app.add_loading_state(
+        LoadingState::new(GameState::Loading).continue_to_state(GameState::StartMenu),
+    )
+    .add_collection_to_loading_state::<_, SpriteAssets>(GameState::Loading)
+    .add_collection_to_loading_state::<_, AudioAssets>(GameState::Loading)
+    .add_collection_to_loading_state::<_, UiAssets>(GameState::Loading);
 
-    app.add_systems(Startup, setup);
+    app.insert_resource(ClearColor(Color::rgb(0., 0., 0.)));
+
+    app.add_systems(
+        OnTransition {
+            from: GameState::Loading,
+            to: GameState::StartMenu,
+        },
+        setup,
+    );
 
     app.run();
 }
 
 /// The setup function
-fn setup(
-    mut commands: Commands,
-    audios: Res<AudioAssets>,
-    mut rapier_configuration: ResMut<RapierConfiguration>,
-) {
+fn setup(mut rapier_configuration: ResMut<RapierConfiguration>) {
     rapier_configuration.gravity = Vec2::ZERO;
-
-    // TODO: Moved here from menu to prevent reloading every time the credits are toggled.
-    //       In reality, we do want this to be respawned when the menu is re-entered,
-    //       just not if the previous state was also a menu state (e.g. Credits).
-    commands.spawn((
-        AudioBundle {
-            source: audios.title_music.clone(),
-            settings: PlaybackSettings {
-                mode: PlaybackMode::Loop,
-                ..default()
-            },
-        },
-        ForState {
-            states: AppState::IN_MENU_STATE.to_vec(),
-        },
-        Name::new("Menu Music"),
-    ));
 }